+++ api/controllers/schemas/ServiceSchema.jsx
@@ -1,10 +1,24 @@
-ServiceSchema = {
-    name: {
-    	label: "Name",
-    	defaultValue: ""
+ComplianceRuleSchema = {
+    service:{
+    	type:Object,
     },
     type: {
-    	label:"Property type",
-    	size:6
-    }
+    	label:"Rule type",
+    	input:"MDSelect",
+    	options:{
+    		items:[
+    			"Contract expiry",				//verify existence of ppm contract
+    			"PPM schedule established",		//verify existence of ppm events
+    			"Service reports on file",		//verify service reports for all ppm items
+    			"PPM event",					//individual event
+    		]
+    	}
+    },
+    name: {
+    	label: "Rule name",
+    	defaultValue: ""
+    },
+
+
+
 }