+++ ui/views/Facility/FacilityViewDetail.jsx
@@ -151,6 +151,12 @@ FacilityViewDetail = React.createClass({
         }
     },
 
+    editService(svc) {
+        Modal.show({
+            content:<div><h1>Edit Service "{svc.name}"</h1></div>
+        });
+    },
+
     render() {
         var facility = this.data.facility;
         var suppliers = this.data.suppliers;
@@ -181,21 +187,37 @@ FacilityViewDetail = React.createClass({
                     :null}
                 </div>
 
+                <div className="title-padding"/>
                 <div className="title-overlay">
-                    <h2>{facility.getName()}</h2>                        
-                    {address?<b>{address}</b>:null}
+                    <div className="row">
+                        <div className="col-md-4">
+                            <div className="facility-title">
+                                <h2>{facility.getName()}</h2>                        
+                                {address?<b>{address}</b>:null}
+                            </div>
+                            {contact?
+                                 <div className="contact-info">
+                                    {contactName?<span className="contact-title">Contact: {contactName}<br/></span>:null}
+                                    <span><i className="fa fa-envelope"></i> {contact.email}<br/></span>
+                                    {contact.phone?<span><i className="fa fa-phone"></i> {contact.phone}<br/></span>:null}
+                                    {contact.phone2?<span><i className="fa fa-phone"></i> {contact.phone2}<br/></span>:null}
+                                </div>
+                            :null}
+
+                        </div>
+                        <div className="col-md-8">
+                            <div className="services-overview">
+                                {facility.servicesRequired?facility.servicesRequired.map((svc,idx)=>{
+                                    return (
+                                        <div className="md-chip" key={idx} onClick={()=>{this.editService(svc)}}><div className="left"/><div className="center"/><div className="right"/><div className="content"><div className="chip-text">{svc.name}</div><div className="chip-close"/></div></div>
+                                    )
+                                }):null}
+                            </div>
+                        </div>
+                    </div>
                 </div>
 
-                {contact?
 
-                     <div className="contact-info">
-                        {contactName?<span className="contact-title">Contact: {contactName}<br/></span>:null}
-                        <span><i className="fa fa-envelope"></i> {contact.email}<br/></span>
-                        {contact.phone?<span><i className="fa fa-phone"></i> {contact.phone}<br/></span>:null}
-                        {contact.phone2?<span><i className="fa fa-phone"></i> {contact.phone2}<br/></span>:null}
-                    </div>
-
-                :null}
 
                 <IpsoTabso tabs={[
                     {